Testing hot config refresh
Once updated and committed (and necessarily pushed, if there is a remote) config files on repo connected to config server, this last one is automatically refreshed.
The client receiving a post on /refresh actuator can have a hot refresh without restarting the application. In this case only Beans with @RefreshScope annotation will be re-injected with properties related Configuration Beans and Values.